добавляет переключение модулей
