+++ js/modules/showScreen.js
@@ -1,0 +1,7 @@
+const showScreen = (page) => {
+  const container = document.querySelector(`.central`);
+  container.innerHTML = ``;
+  container.appendChild(page.element);
+};
+
+export default showScreen;
