+++ js/templates/game-one-template.js
@@ -1,0 +1,42 @@
+const gameOneString = `<div class="game">
+    <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
+    <form class="game__content">
+      <div class="game__option">
+        <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
+        <label class="game__answer game__answer--photo">
+          <input name="question1" type="radio" value="photo">
+          <span>Фото</span>
+        </label>
+        <label class="game__answer game__answer--paint">
+          <input name="question1" type="radio" value="paint">
+          <span>Рисунок</span>
+        </label>
+      </div>
+      <div class="game__option">
+        <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
+        <label class="game__answer  game__answer--photo">
+          <input name="question2" type="radio" value="photo">
+          <span>Фото</span>
+        </label>
+        <label class="game__answer  game__answer--paint">
+          <input name="question2" type="radio" value="paint">
+          <span>Рисунок</span>
+        </label>
+      </div>
+    </form>
+    <div class="stats">
+      <ul class="stats">
+        <li class="stats__result stats__result--wrong"></li>
+        <li class="stats__result stats__result--slow"></li>
+        <li class="stats__result stats__result--fast"></li>
+        <li class="stats__result stats__result--correct"></li>
+        <li class="stats__result stats__result--unknown"></li>
+        <li class="stats__result stats__result--unknown"></li>
+        <li class="stats__result stats__result--unknown"></li>
+        <li class="stats__result stats__result--unknown"></li>
+        <li class="stats__result stats__result--unknown"></li>
+        <li class="stats__result stats__result--unknown"></li>
+      </ul>
+    </div>
+  </div>`;
+export default gameOneString;
