+++ js/templates/stats-template.js
@@ -1,16 +1,4 @@
-import createEl from './createEl.js';
-import showScreen from './showScreen.js';
-import moduleGreetingElement from './greeting.js';
-
-const statsEl = `<header class="header">
-    <div class="header__back">
-      <span class="back">
-        <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
-        <img src="img/logo_small.png" width="101" height="44">
-      </span>
-    </div>
-  </header>
-  <div class="result">
+const statsString = `<div class="result">
     <h1>Победа!</h1>
     <table class="result__table">
       <tr>
@@ -109,21 +97,6 @@ const statsEl = `<header class="header">
         <td colspan="5" class="result__total  result__total--final">950</td>
       </tr>
     </table>
-  </div>
-  <footer class="footer">
-    <a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
-    <span class="footer__made-in">Сделано в <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a> &copy; 2016</span>
-    <div class="footer__social-links">
-      <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
-      <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
-      <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
-      <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
-    </div>
-  </footer>`;
+  </div>`;
 
-const moduleStatsElement = createEl(statsEl);
-const backButton = moduleStatsElement.querySelector(`.header__back`);
-backButton.addEventListener(`click`, () => {
-  showScreen(moduleGreetingElement);
-});
-export default moduleStatsElement;
+export default statsString;
